feat: remove drawer and text-to-speech with expo on articles
