+++ apps/native/app/(app)/(tabs)/trails/[id].tsx
