+++ apps/native/app/(app)/(tabs)/trails/_layout.tsx
