+++ apps/native/app/(app)/(tabs)/trails/index.tsx
