+++ apps/native/package.json
@@ -37,6 +37,7 @@
     "expo-navigation-bar": "~5.0.8",
     "expo-router": "~6.0.10",
     "expo-secure-store": "~15.0.7",
+    "expo-speech": "~14.0.7",
     "expo-status-bar": "~3.0.8",
     "expo-system-ui": "~6.0.7",
     "expo-web-browser": "^15.0.8",
